See merge request benkuly/trixnity!134

clientserverapi: add missing endpoints

Closes #97

closes #97

+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/media/MediaService.kt
@@ -79,8 +79,8 @@ class MediaService(
 
     suspend fun getThumbnail(
         mxcUri: String,
-        width: UInt,
-        height: UInt,
+        width: Long,
+        height: Long,
         method: ThumbnailResizingMethod = CROP,
         progress: MutableStateFlow<FileTransferProgress?>? = null
     ): Result<ByteArray> = kotlin.runCatching {
