+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/keys/ClaimKeys.kt
@@ -4,8 +4,8 @@ import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonElement
-import net.folivo.trixnity.core.HttpMethodType.POST
 import net.folivo.trixnity.core.HttpMethod
+import net.folivo.trixnity.core.HttpMethodType.POST
 import net.folivo.trixnity.core.MatrixEndpoint
 import net.folivo.trixnity.core.model.UserId
 import net.folivo.trixnity.core.model.keys.KeyAlgorithm
@@ -22,7 +22,7 @@ data class ClaimKeys(
         @SerialName("one_time_keys")
         val oneTimeKeys: Map<UserId, Map<String, KeyAlgorithm>>,
         @SerialName("timeout")
-        val timeout: Int?,
+        val timeout: Long?,
     )
 
     @Serializable
