+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/rooms/SetTyping.kt
@@ -3,8 +3,8 @@ package net.folivo.trixnity.clientserverapi.model.rooms
 import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import net.folivo.trixnity.core.HttpMethodType.PUT
 import net.folivo.trixnity.core.HttpMethod
+import net.folivo.trixnity.core.HttpMethodType.PUT
 import net.folivo.trixnity.core.MatrixEndpoint
 import net.folivo.trixnity.core.model.RoomId
 import net.folivo.trixnity.core.model.UserId
@@ -20,6 +20,6 @@ data class SetTyping(
     @Serializable
     data class Request(
         @SerialName("typing") val typing: Boolean,
-        @SerialName("timeout") val timeout: Int? = null,
+        @SerialName("timeout") val timeout: Long? = null,
     )
 }
